Removed unneded codecs
